M.Chunk.Code

Description

Encode and decode paletted containers for block states and biomes.

Documentation

data ChunkSection c m

a chunk section where c is the numeric type for block states and m is the same for biomes

Constructors

ChunkSection

Fields

csnonempty :: !Int16

number of non-air blocks (tracked for optimization)

csblockstates :: !(Vector c)

block states (4,096 entries; 16x16x16, access [y][z][x])

csbiomes :: !(Vector m)

biomes (64 entries; 4x4x4, access [y][z][x])

data ChunkSectionEncoding

encoding configuration for ChunkSection

Constructors

ChunkSectionEncoding

Fields

cseblockstates :: !Int

number of possible block states

csebiomes :: !Int

number of possible biomes

mkcscodec

Arguments

:: forall m c (st :: ZeroBitType) r. (Unbox m, Unbox c, FiniteBits m, FiniteBits c, Integral m, Integral c)

=> ChunkSectionEncoding
encoding settings

-> (ChunkSection c m -> Builder, Parser st r (ChunkSection c m))
a pair of an encoder and a decoder, respectively

create a codec for ChunkSections using the provided settings