M.Collision.Internal.Face

Description

Provides utilities for sampling points on cuboid faces, used in collision detection. The module focuses on efficient face point generation for AABB collision testing.

Documentation

facepoints

Arguments

:: V3 Int
cuboid dimensions

-> V3 Int
direction (signum vector)

-> [V3 Int]
points on the faces from the cuboid's (0, 0, 0) corner

retrieve the points on the faces of a cuboid normal to a certain direction (sig: {-1, 0, 1}) from said cuboid

dbgcountfacepoints

Arguments

:: V3 Int
cuboid dimensions

-> V3 Int
direction (signum vector)

-> Int
number of points sampled by facepoints

predict the number of points sampled by facepoints

dbgdesmos :: Show a => [V3 a] -> String

print a list of vectors in a format that can be copy-pasted into Desmos