M.Collision.Pure

Description

Core collision detection primitives and algorithms in pure form. Provides AABB collision testing, shape interfaces, and hit detection utilities.

Documentation

class Shape (s :: Type -> Type) where

an AABB type class used for collision detection and resolution

Minimal complete definition

intersecting, crossing, hitting, translate, corners, tomanyboxes

Methods

intersecting :: (Fractional a, Ord a) => s a -> s a -> Bool

check if two shapes intersect

crossing :: RealFloat a => V3 a -> V3 a -> s a -> Hit a

check if a ray will hit the shape and return the hit data

hitting :: RealFloat a => V3 a -> s a -> s a -> Hit a

check if the first shape will collide into the second shape if it moves with the given displacement

translate :: Num a => V3 a -> s a -> s a

translate the shape by the given displacement

corners :: (Fractional a, Ord a) => s a -> V2 (V3 a)

the locations of the lower and higher corners of the shape respectively

tomanyboxes :: s a -> ManyBoxes [] a

convert a Shape to a ManyBoxes of Boxes with a list container, which is a canonical form for ManyBoxes

scenter :: (Fractional a, Ord a) => s a -> V3 a

the center of the shape

sdimensions :: (Fractional a, Ord a) => s a -> V3 a

the dimensions of the shape

data SomeShape1 a

existential Shape type but where numeric type is erased

see also: castshape1

Constructors

(Typeable (s a), Show (s a), Shape s) => SomeShape1 (s a)

data Hit a

a collision resolution data type

no hit is represented by a hit at infinity (other fields are unspecified)

Constructors

Hit

Fields

hittime :: !a

proportion of move completed in [0, 1]

hitwhere :: !(V3 a)

the point of collision

if you're using Box, this is the center of the box

hitnorm :: !(V3 a)

normal vector of the surface hit

a signum vector, so each component is either -1, 0, or 1

newtype Hit' a

internal newtype used with Min to find the closest hit

Constructors

Hit'

Fields

unHit' :: Hit a

data Box a

a box in 3D space, located either relatively or absolutely

Constructors

Box

Fields

dimensions :: !(V3 a)

the dimensions of the box

center :: !(V3 a)

the center of the box

Bundled Patterns

pattern Box' :: Fractional a => V3 a -> V3 a -> Box a
bidrectional pattern for Box but with corner locations (low to high)

you can use the locorner' and hicorner' patterns to extract the corners, respectively

newtype ManyBoxes (f :: Type -> Type) a

a newtype over a Foldable Functor container of Boxes

the low and high corners are those of the smallest bounding box

Constructors

ManyBoxes (f (Box a))

type ManyBoxes_ a = ManyBoxes [] a

a type alias for a list of Boxes (canonical form for ManyBoxes)

_dimensions :: forall a f. Functor f => (V3 a -> f (V3 a)) -> Box a -> f (Box a)

Lens for the dimensions of the box

_center :: forall a f. Functor f => (V3 a -> f (V3 a)) -> Box a -> f (Box a)

Lens for the center of the box

_lcorner :: Fractional a => Lens' (Box a) (V3 a)

Lens for the lower corner of the box

_hcorner :: Fractional a => Lens' (Box a) (V3 a)

Lens for the higher corner of the box

hitin01 :: (Num a, Ord a) => Hit a -> Bool

check if the hit time is in [0, 1]

infhit :: Fractional a => Hit a

a hit at infinity

boxfromcorners

Arguments

:: Fractional a

=> V3 a
low corner

-> V3 a
high corner

-> Box a
the box

a box from the low and high corners

castshape1 :: Typeable b => SomeShape1 a -> Maybe b

cast a SomeShape1 to a specific type

boxzero :: Num a => Box a

a box with zero dimensions and center

hicorner :: Fractional a => Box a -> V3 a

the location of the higher corner of the box

hicorner' :: Fractional a => Box a -> V3 a

locorner :: Fractional a => Box a -> V3 a

the location of the lower corner of the box

locorner' :: Fractional a => Box a -> V3 a

shicorner :: (Shape s, Fractional a, Ord a) => s a -> V3 a

the upper corner of a shape

slocorner :: (Shape s, Fractional a, Ord a) => s a -> V3 a

the lower corner of a shape