M.Crypto

Description

Provides cryptographic functionality required by the Minecraft protocol, including AES encryption/decryption, RSA operations, and server-specific hashing utilities.

AES

data AES (mode :: Mode)

encryption/decryption context for AES (AES-128-CFB8)

data Mode

encryption/decryption mode

Constructors

Encrypt

Decrypt

aesnew :: AESClass mode => ByteString -> IO (AES mode)

create a new AES context

aesupdate :: AESClass mode => AES mode -> ByteString -> IO ByteString

either encrypt or decrypt a message

RSA

data RSA

RSA pkey

rsanew :: Int -> IO RSA

create a new RSA pkey

rsaup :: RSAClass mode => RSA -> ByteString -> IO ByteString

either encrypt or decrypt a message

rsaspki :: RSA -> IO ByteString

write the SubjectPublicKeyInfo to a ByteString

Hash

hashservnam

Arguments

:: ByteString
server id

-> ByteString
shared secret

-> ByteString
verify token

-> IO ByteString
SHA1 hash (20 bytes)

generate a SHA1 hash of the server name

General

newtype Error

error type

Constructors

Error String

class AESClass (mode :: Mode)

Minimal complete definition

aesnew_, aesupdate_

class RSAClass (mode :: Mode)

Minimal complete definition

rsaup_

Re-exports

data ByteString

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Char8 it can be interpreted as containing 8-bit characters.