M.IO

Description

Provides high-level IO operations and types for handling Minecraft protocol connections, including socket management and data streaming.

Documentation

data Connection

a connection to either a server or a client

Constructors

Connection

Fields

cxkey :: TVar (Maybe ByteString)

encryption key, AES-128-CFB8; doubles as IV

cxcompth :: TVar Int

compression threshold; negative = off, non-negative = on with threshold

cxinput :: InputStream Uninterpreted

input stream

cxoutput :: OutputStream Uninterpreted

output stream

data Uninterpreted

uninterpreted packet

Constructors

Uninterpreted

Fields

pkcode :: !Word8

pkdata :: !ByteString

define communication effects

module M.IO.Internal.EffectTypes

withtalkingserver

Arguments

:: forall (es :: [Effect]) a. (IOE :> es, State ParserState :> es, Concurrent :> es, NonDet :> es)

=> Maybe String
host (Nothing = all interfaces)

-> String
port

-> Eff (Talking ': es) a
per-connection handler

-> Eff es a
final result

run server accepting multiple connections

withcxfromsocket :: MonadUnliftIO m => Socket -> (Connection -> m a) -> m a

create a connection from a socket