M.IO.Internal.Datagram

Description

Internal module for handling low-level packet parsing and building, including uninterpreted packets and stream transformations.

Types

data Uninterpreted

uninterpreted packet

Constructors

Uninterpreted

Fields

pkcode :: !Word8

pkdata :: !ByteString

data EOF

end of input

Constructors

EOF

Streams

makepacketstreami

Arguments

:: TVar Int
compression threshold reference (negative = off, non-negative = on with threshold)

-> InputStream ByteString
input stream

-> IO (InputStream Uninterpreted)
stream of uninterpreted packets

make a stream of uninterpreted packets

makepacketstreamo :: TVar Int -> OutputStream ByteString -> IO (OutputStream Uninterpreted)

make an output stream of uninterpreted packets

makedecrypting

Arguments

:: TVar (ByteString -> IO ByteString)
decryptor

-> InputStream ByteString
input stream

-> IO (InputStream ByteString)
new input stream

register an octet streaming decryptor to an input stream

makeencrypting

Arguments

:: TVar (ByteString -> IO ByteString)
encryptor

-> OutputStream ByteString
output stream

-> IO (OutputStream ByteString)
new output stream

register an octet stremaing encryptor to an output stream