M.IO.Internal.EffectSocket

Description

Implements the interpretation of the Talking effect in terms of socket connections, providing both client and server capabilities.

Documentation

data SocketTalkingError

error in communication

Constructors

UnknownCode Direction TypeRep

Mismatch Direction TypeRep TypeRep

withtalkingserver

Arguments

:: forall (es :: [Effect]) a. (IOE :> es, State ParserState :> es, Concurrent :> es, NonDet :> es)

=> Maybe String
host (Nothing = all interfaces)

-> String
port

-> Eff (Talking ': es) a
per-connection handler

-> Eff es a
final result

run server accepting multiple connections