M.IO.Internal.Read

Description

This module provides parsing utilities for reading structured data from streams.

Basic parsing

parseio

Arguments

:: Exception e

=> r
state

-> Int
int value

-> InputStream ByteString
input stream

-> ParserIO r e a
parser

-> IO a
result

Parse from a stream. Automatically handles chunked input by concatenating chunks until a complete parse succeeds.

May throw: * The parser's error type e * IOError "parseio: unexpected end of input"

parseio0

Arguments

:: Exception e

=> InputStream ByteString
input stream

-> ParserIO () e a
parser

-> IO a
result

parse from a stream (with no state and int value of 0). see also: parseio

Lifted versions

parseiolift

Arguments

:: (MonadIO m, Exception e)

=> r
state

-> Int
int value

-> InputStream ByteString
input stream

-> ParserIO r e a
parser

-> m a
result

lifted version of parseio

parseio0lift

Arguments

:: (MonadIO m, Exception e)

=> InputStream ByteString
input stream

-> ParserIO () e a
parser

-> m a
result

lifted version of parseio0