M.IO.Internal.Socket

Description

Implements reliable duplex stream connections for the Java Minecraft protocol, handling encryption and compression.

Documentation

data Connection

a connection to either a server or a client

Constructors

Connection

Fields

cxkey :: TVar (Maybe ByteString)

encryption key, AES-128-CFB8; doubles as IV

cxcompth :: TVar Int

compression threshold; negative = off, non-negative = on with threshold

cxinput :: InputStream Uninterpreted

input stream

cxoutput :: OutputStream Uninterpreted

output stream

withcxfromsocket :: MonadUnliftIO m => Socket -> (Connection -> m a) -> m a

create a connection from a socket