M.LEB

Description

Implements LEB128 (Little-Endian Base 128) variable-length encoding for arbitrary finite-bit integers, used in the Minecraft protocol.

Documentation

newtype LEB a

a LEB128 (unsigned) encoded integer. the integer itself may or may not be signed

Constructors

LEB

Fields

getleb :: a

type VarInt = LEB Int32

VarInt

type VarLong = LEB Int64

VarLong

decodeleb

Arguments

:: (Monad m, FiniteBits a, Num a)

=> m Word8
accept any Word8 value

-> m (LEB a)
LEB128 encoded integer

decode an unsigned LEB128 encoded integer

the actual integer may or may not be signed; the encoding itself is "unsigned"

encodeleb :: (FiniteBits a, Integral a) => LEB a -> Builder

encode an LEB128 encoded integer into a Builder