M.Misc

Description

Defines common types used across the Minecraft protocol implementation, including teleportation flags and sound events.

Documentation

data TeleportFlags

flags for teleporting an entity

16+ bits

Constructors

TeleportFlags

Fields

tprelx :: Bool

tprely :: Bool

tprelz :: Bool

tprelyaw :: Bool

tprelpitch :: Bool

tprelvelx :: Bool

tprelvely :: Bool

tprelvelz :: Bool

tprelvelyaw :: Bool

tprotvelfirst :: Bool

from The Minecraft Wiki: "Rotate velocity according to the change in rotation, before applying the velocity change in this packet. Combining this with absolute rotation works as expected—the difference in rotation is still used."

data SoundEvent

Constructors

SoundEvent

Fields

soundname :: Text

fixedrange :: Maybe Float