M.NBT.Internal.JS

Description

Implements Java's Modified UTF-8 (CESU-8) encoding and decoding for string handling in NBT format.

See: https://docs.oracle.com/javase/8/docs/api/java/io/DataInput.html#modified-utf-8

Documentation

newtype JS

Java's CESU-8 encoding/decoding

this newtype is purely for modulation of encoding and decoding. it is not intended to be used directly in the public API

use textascesu8 and cesu8astext to convert between Text and ByteString

see:

https://en.wikipedia.org/wiki/UTF-8#CESU-8

https://docs.oracle.com/en/java/javase/18/docs/api/java.base/java/io/DataInput.html#modified-utf-8

Constructors

JS

Fields

getjs :: Text

textascesu8 :: Text -> ByteString

encode Text into CESU-8 ByteString

cesu8astext :: ByteString -> Maybe Text

decode CESU-8 ByteString into Text

tocesu8 :: JS -> ByteString

encode text to CESU-8

tocesu8p :: JS -> Builder

encode text to CESU-8 (Builder version)

fromcesu8 :: ByteString -> Maybe JS

decode CESU-8 encoded text

fromcesu8p :: forall (st :: ZeroBitType) r. Parser st r JS

decode CESU-8 encoded text (Parser version)