M.Pack.Internal.Linear

Description

Provides Pack and Unpack instances for linear algebra types from the linear package, including vectors and complex numbers.

Orphan instances

Pack a => Pack (Complex a)

Instance details

Methods

pack :: Complex a -> Builder

Pack a => Pack (Quaternion a)

Instance details

Methods

pack :: Quaternion a -> Builder

Pack a => Pack (V0 a)

Instance details

Methods

pack :: V0 a -> Builder

Pack a => Pack (V1 a)

Instance details

Methods

pack :: V1 a -> Builder

Pack a => Pack (V2 a)

Instance details

Methods

pack :: V2 a -> Builder

Pack a => Pack (V3 a)

Instance details

Methods

pack :: V3 a -> Builder

Pack a => Pack (V4 a)

Instance details

Methods

pack :: V4 a -> Builder

Unpack a => Unpack (Complex a)

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r (Complex a)

Unpack a => Unpack (Quaternion a)

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r (Quaternion a)

Unpack a => Unpack (V0 a)

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r (V0 a)

Unpack a => Unpack (V1 a)

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r (V1 a)

Unpack a => Unpack (V2 a)

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r (V2 a)

Unpack a => Unpack (V3 a)

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r (V3 a)

Unpack a => Unpack (V4 a)

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r (V4 a)

Pack a => Pack (V n a)

Instance details

Methods

pack :: V n a -> Builder

(Unpack a, Dim n) => Unpack (V n a)

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r (V n a)