M.Pack.Internal.Newtypes

Description

Defines newtype wrappers that control how values are packed and unpacked, including enum indices, fixed-point numbers, angles, and identifiers.

Documentation

newtype EnumIndex (i :: k) a

represent any Enum type using a zero-based index

Constructors

EnumIndex

Fields

enumindex :: a

newtype Fixed' (i :: k) (r :: k1) f

use an integer type i for serialization of a fixed-point number with resolution r (see HasResolution, Fixed), with underlying numeric representation f

Constructors

Fixed'

Fields

unfixed' :: f

newtype Int8Angle

a signed angle; divides the circle into 256 parts

Constructors

Int8Angle

Fields

int8angle :: Int8

newtype Identifier

an "identifier" string, used for names, tags, etc.

example: minecraft:stone or stone/stone

validation only happens when unpacking (unpack)

Constructors

Identifier

Fields

identifier :: Text

newtype IDorX a

unresolved value; either an ID or an inline value

Constructors

IDorX

Fields

idorvalue :: Either Int32 a

VarInt ID (NOT +1) or inline value

newtype IDSet

potentially unresolved ID set; either an identifier for its location or an inline set of IDs

Constructors

IDSet

Fields

setnameorids :: Either Text (Vector Int32)

name of ID set or inline set

newtype TakeRest

a newtype wrapper over ByteString; not length-prefixed

Constructors

TakeRest

Fields

gettakerest :: ByteString

newtype PackFoldableVI (f :: k -> Type) (a :: k)

General Pack instance provider for Foldables

length-prefixed (VarInt), strict left fold

Constructors

PackFoldableVI

Fields

getpackfoldablevi :: f a

newtype PackFoldable0 (f :: k -> Type) (a :: k)

General Pack instance provider for Foldables

no length prefix, strict left fold

Constructors

PackFoldable0

Fields

getpackfoldable0 :: f a

newtype UnpackRepresentable0 (f :: k -> Type) (a :: k)

General Unpack instance provider for Representables that are also Traversable

Constructors

UnpackRepresentable0

Fields

getunpackrepresentable0 :: f a

degtoi8angle :: RealFrac a => a -> Int8Angle

convert degrees to Int8Angle

i8angledeg :: RealFrac a => Int8Angle -> a

convert Int8Angle to degrees