M.Pack.Internal.Num

Description

Implements Pack and Unpack instances for numeric types and Bool, including integral types, floating point numbers, and LEB128 encoding.

Documentation

packleb32 :: Integral a => a -> Builder

pack as an LEB128-encoded Int32

unpackleb32 :: forall a (st :: ZeroBitType) r. Integral a => Parser st r a

unpack from an LEB128-encoded Int32

packfi :: (Integral a, Pack a, Integral b) => b -> Builder

pack b in the format of a

unpackfi :: forall a b (st :: ZeroBitType) r. (Integral a, Unpack a, Integral b) => Parser st r b

unpack b from a number in the format of a

guardnat :: forall a (st :: ZeroBitType) r. (Num a, Ord a, Show a) => String -> a -> Parser st r a

ensure the number is non-negative

Orphan instances

Pack Int16

Instance details

Methods

pack :: Int16 -> Builder

Pack Int32

Instance details

Methods

pack :: Int32 -> Builder

Pack Int64

Instance details

Methods

pack :: Int64 -> Builder

Pack Int8

Instance details

Methods

pack :: Int8 -> Builder

Pack Word16

Instance details

Methods

pack :: Word16 -> Builder

Pack Word32

Instance details

Methods

pack :: Word32 -> Builder

Pack Word64

Instance details

Methods

pack :: Word64 -> Builder

Pack Word8

Instance details

Methods

pack :: Word8 -> Builder

Pack Bool

Instance details

Methods

pack :: Bool -> Builder

Pack Double

Instance details

Methods

pack :: Double -> Builder

Pack Float

Instance details

Methods

pack :: Float -> Builder

Unpack Int16

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r Int16

Unpack Int32

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r Int32

Unpack Int64

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r Int64

Unpack Int8

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r Int8

Unpack Word16

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r Word16

Unpack Word32

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r Word32

Unpack Word64

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r Word64

Unpack Word8

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r Word8

Unpack Bool

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r Bool

Unpack Double

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r Double

Unpack Float

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r Float

(FiniteBits a, Integral a) => Pack (LEB a)

Instance details

Methods

pack :: LEB a -> Builder

(FiniteBits a, Num a) => Unpack (LEB a)

Instance details

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r (LEB a)