M.Pack.Internal.TH

Description

Provides Template Haskell machinery for automatically deriving Pack and Unpack instances, including support for shadowing and proper derivation.

Documentation

borrowderivepackunpack :: RunUserCoercion -> Q [Dec]

shadow-derive Pack and Unpack instances for a type

properderivepackunpack :: Name -> Q [Dec]

borrowderivenothing :: RunUserCoercion -> Q [Dec]

literally do nothing