M.Pack.Internal.Types

Description

Defines the core type system for binary serialization and deserialization, including Pack and Unpack typeclasses and related utilities.

Documentation

class Pack a where

produce a Builder from a value

Minimal complete definition

Nothing

Methods

pack :: a -> Builder

produce a Builder from a value

default pack :: (Generic a, GPack (Rep a)) => a -> Builder

data SomePack

existential Pack container

Constructors

(Typeable a, Pack a, Show a) => SomePack a

class Unpack a where

retrieve a Parser for a type

Minimal complete definition

Nothing

Methods

unpack :: forall (st :: ZeroBitType) r. Parser st r a

retrieve a Parser for a type

default unpack :: forall (st :: ZeroBitType) r. (Generic a, GUnpack (Rep a)) => Parser st r a

data SomeUnpack

existential Unpack container

to unpack this, try using unpacksome with a type application

Constructors

(Typeable a, Unpack a, Show a) => SomeUnpack a

type Parser (st :: ZeroBitType) r = ParserT st r ParseError

our parser type

type Result = Result ParseError

our parse result

newtype ParseError

our parser error type

Constructors

ParseError

Fields

showparseerror :: String

parsepure :: (forall (st :: ZeroBitType). Parser st r a) -> r -> Int -> ByteString -> Result a

run a pure parser efficiently (by inlining)

parsepure0 :: (forall (st :: ZeroBitType). Parser st () a) -> ByteString -> Result a

run a pure parser with no state efficiently (by inlining)

castsomepack :: Typeable a => SomePack -> Maybe a

cast a SomePack to a type

castsomeunpack :: Typeable a => SomeUnpack -> Maybe a

cast a SomeUnpack to a type

unpacksome :: forall a (st :: ZeroBitType) r. (Typeable a, Unpack a, Show a) => Parser st r SomeUnpack

use a type application to retrieve a SomeUnpack