M.V769.C

Description

Defines packet structures for the configuration state of protocol version 769, including client and server capabilities negotiation, resource pack management, and server feature flags.

Documentation

data ResourcePackResponse

Constructors

ResourcePackResponse

Fields

status :: VarInt

data Pong

Constructors

Pong

Fields

id :: Int64

data AcknowledgeFinishConfiguration

Constructors

AcknowledgeFinishConfiguration

data CookieResponse

Constructors

CookieResponse

Fields

key :: Text

value :: Text

data ClientInformationConfiguration__

Constructors

ClientInformationConfiguration__

Fields

locale__ :: Text

viewdistance__ :: Int8

chatmode__ :: EnumIndex VarInt ChatMode

chatcolors__ :: Bool

displayedskinparts__ :: Bitwise Word8 DisplayedSkinParts

mainhandright__ :: Bool

enabletextfiltering__ :: Bool

allowserverlistings__ :: Bool

particlestatus__ :: EnumIndex VarInt ParticleStatus

data ClientInformationConfiguration

Constructors

ClientInformationConfiguration

Fields

locale :: Text

viewdistance :: Int8

chatmode :: ChatMode

chatcolors :: Bool

displayedskinparts :: DisplayedSkinParts

mainhandright :: Bool

enabletextfiltering :: Bool

allowserverlistings :: Bool

particlestatus :: ParticleStatus

data ServerLinks__

Constructors

ServerLinks__

Fields

links__ :: Vector Identifier

data ServerLinks

Constructors

ServerLinks

Fields

links :: Vector Text

data CustomReportDetails__

Constructors

CustomReportDetails__

Fields

details__ :: TakeRest

data CustomReportDetails

Constructors

CustomReportDetails

Fields

details :: ByteString

data KnownPacks__

Constructors

KnownPacks__

Fields

packs__ :: TakeRest

data KnownPacks

Constructors

KnownPacks

Fields

packs :: ByteString

data UpdateTags__

Constructors

UpdateTags__

Fields

tags__ :: TakeRest

data UpdateTags

Constructors

UpdateTags

Fields

tags :: ByteString

data FeatureFlags__

Constructors

FeatureFlags__

Fields

flags__ :: Vector Identifier

data FeatureFlags

Constructors

FeatureFlags

Fields

flags :: Vector Text

data Transfer

Constructors

Transfer

Fields

serveraddress :: Text

serverport :: Word16

data StoreCookie

Constructors

StoreCookie

Fields

key :: Text

value :: Text

data RemoveResourcePack

Constructors

RemoveResourcePack

data RegistryData

Constructors

RegistryData

Fields

data_ :: ByteString

data ResetChat

Constructors

ResetChat

data FinishConfiguration

Constructors

FinishConfiguration

data Disconnect

Constructors

Disconnect

Fields

reason :: Text

data CookieRequest

Constructors

CookieRequest

Fields

key :: Text