M.V769.H

Description

Implements packet structures for the initial handshake phase of protocol version 769, where clients establish basic connection parameters including protocol version and intended next state.

Documentation

data HandshakePacket__

Constructors

HandshakePacket__

Fields

protocolversion__ :: VarInt

serveraddress__ :: Text

serverport__ :: Word16

nextstate__ :: VarInt

data HandshakePacket

Constructors

HandshakePacket

Fields

protocolversion :: Int32

serveraddress :: Text

serverport :: Word16

nextstate :: Int32