M.V769.L

Description

Defines packet structures for the login phase of protocol version 769, handling authentication, encryption setup, and session establishment between client and server.

Documentation

data EncryptionResponse

Constructors

EncryptionResponse

Fields

sharedsecret :: ByteString

verifytoken :: ByteString

data LoginAcknowledged

Constructors

LoginAcknowledged

data LoginStart

Constructors

LoginStart

Fields

name :: Text

uuid :: UUID

data LoginPluginResponse__

Constructors

LoginPluginResponse__

Fields

messageid__ :: VarInt

successful__ :: Bool

data___ :: Maybe ByteString

data LoginPluginResponse

Constructors

LoginPluginResponse

Fields

messageid :: Int32

successful :: Bool

data_ :: Maybe ByteString

data LoginPluginRequest__

Constructors

LoginPluginRequest__

Fields

messageid__ :: LEB Int32

channel__ :: Text

data___ :: TakeRest

data LoginPluginRequest

Constructors

LoginPluginRequest

Fields

messageid :: Int32

channel :: Text

data_ :: ByteString

data LoginSuccess_Property

Constructors

LoginSuccess_Property

Fields

name :: Text

value :: Text

signature :: Maybe Text

data LoginSuccess

Constructors

LoginSuccess

Fields

uuid :: UUID

username :: Text

properties :: Vector LoginSuccess_Property

data EncryptionRequest

Constructors

EncryptionRequest

Fields

serverid :: Text

publickey :: ByteString

verifytoken :: ByteString

shouldauthenticate :: Bool

data LoginDisconnect

Constructors

LoginDisconnect

Fields

reason :: TextComponent