M.V769.S

Description

Defines packet structures for the status state of protocol version 769, handling server list ping functionality including server status queries and latency checks.

Documentation

data PongResponse

Constructors

PongResponse

Fields

payload :: Int64

data StatusRequest

Constructors

StatusRequest

data StatusResponse

Constructors

StatusResponse

Fields

jsonresponse :: Text